backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
python-dotenv==1.0.1
langchain==0.1.13
langchain-openai==0.1.1
playwright==1.42.0
google-api-python-client==2.122.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
openpyxl==3.1.2
pandas==2.2.1
python-multipart==0.0.9
PyPDF2==3.0.1
